datapath: Move proxy_hairpin to a new file
Refactor to prepare for a bug fix in the following commit.
